+++ src/org/ojalgo/optimisation/convex/ConvexSolver.java
@@ -254,6 +254,7 @@ public abstract class ConvexSolver extends GenericSolver implements UpdatableSol
         }
 
     }
+
     public static final class ModelIntegration extends ExpressionsBasedModel.Integration<ConvexSolver> {
 
         public ConvexSolver build(final ExpressionsBasedModel model) {
@@ -278,6 +279,7 @@ public abstract class ConvexSolver extends GenericSolver implements UpdatableSol
 
     private static final String Q_NOT_POSITIVE_SEMIDEFINITE = "Q not positive semidefinite!";
     private static final String Q_NOT_SYMMETRIC = "Q not symmetric!";
+    private static final double SMALL_DIAGONAL_FACTOR = RELATIVELY_SMALL + MACHINE_EPSILON;
 
     public static void copy(final ExpressionsBasedModel sourceModel, final ConvexSolver.Builder destinationBuilder) {
 
@@ -632,7 +634,7 @@ public abstract class ConvexSolver extends GenericSolver implements UpdatableSol
         }
 
         if (!mySolverQ.compute(matrixQ)) {
-            matrixQ.modifyDiagonal(ADD.by(RELATIVELY_SMALL * matrixQ.aggregateAll(Aggregator.LARGEST)));
+            matrixQ.modifyDiagonal(ADD.by(SMALL_DIAGONAL_FACTOR * matrixQ.aggregateAll(Aggregator.LARGEST)));
             mySolverQ.compute(matrixQ);
         }