+++ test/org/ojalgo/optimisation/ModelFileMPS.java
@@ -44,6 +44,9 @@ public interface ModelFileMPS {
     static void assertValues(final ExpressionsBasedModel model, final BigDecimal expMinVal, final BigDecimal expMaxVal, final Map<String, BigDecimal> solution,
             final NumberContext precision) {
 
+        //        model.options.debug(Optimisation.Solver.class);
+        //        model.options.mip_defer = 0.0;
+
         TestUtils.assertTrue(model.validate());
 
         if (expMinVal != null) {
@@ -52,9 +55,9 @@ public interface ModelFileMPS {
 
             TestUtils.assertStateNotLessThanFeasible(result);
 
-            TestUtils.assertEquals(expMinVal, result.getValue(), precision);
-
             TestUtils.assertTrue(ModelFileMPS.SOLUTION_NOT_VALID, model.validate(result, precision));
+
+            TestUtils.assertEquals(expMinVal, result.getValue(), precision);
         }
 
         if (expMaxVal != null) {
@@ -63,9 +66,9 @@ public interface ModelFileMPS {
 
             TestUtils.assertStateNotLessThanFeasible(result);
 
-            TestUtils.assertEquals(expMaxVal, result.getValue(), precision);
-
             TestUtils.assertTrue(ModelFileMPS.SOLUTION_NOT_VALID, model.validate(result, precision));
+
+            TestUtils.assertEquals(expMaxVal, result.getValue(), precision);
         }
 
         if (solution != null) {